Fix pipeline stagnation by tracking 'date' in version
In Concourse, when a resource configuration changes its
version history is wiped and a new initial version is created.

Because we were tracking only commit 'ref', if the autopilot
repo hadn't changed the new version created would be identical
to the previous one and no new build would be triggered.

By tracking the 'date' in versions we ensure that whenever one
of these situations happen a new build will be triggered:
- The resource configuration changes
- The resource_type container image is updated

This ensures that the pipelines are always fresh and reflect
any changes in the configuration as soon as they happen.

It also allows you to track these configuration changes by
simply looking at the 'date' field in each build history.