Pave the way for multiheaded TargetConfigs
Currently,to work with managed Concourse instances for which we
don't have access to the 'main' team we are forced to define a
different TargetConfig file for each team we want to autopilot.

Reason is that autopilot can't generate multiple pipelines from
a single TargetConfig. And Concourse forbids pipelines outside
the 'main' team to use the set_pipeline step specifying a team.

A multiheaded TargetConfig file will generate one pipeline for
each team in its packs section. Allowing us to work seamlessly
whether we use a fully-owned or 3rd-party managed Concourse